# packages/api-client/openapi.json
{
...
  "info": {
    "contact": {},
    "description": "REST API for the Hà Tĩnh immersive tourism platform foundation.",
    "title": "Hà Tĩnh Immersive API",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/scenes": {
      "post": {
        "operationId": "createScene",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "altitude": {
                    "nullable": true,
                    "type": "number"
                  },
                  "destinationId": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "geoPoint": {
                    "properties": {
                      "latitude": {
                        "example": 18.3421,
                        "type": "number"
                      },
                      "longitude": {
                        "example": 105.9032,
                        "type": "number"
                      }
                    },
                    "required": ["latitude", "longitude"],
                    "type": "object"
                  },
                  "id": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "initialFov": {
                    "example": 90,
                    "type": "number"
                  },
                  "initialHeading": {
                    "example": 0,
                    "type": "number"
                  },
                  "initialPitch": {
                    "example": 0,
                    "type": "number"
                  },
                  "name": {
                    "example": "Cổng vào khu di tích",
                    "type": "string"
                  },
                  "panoramaAssetId": {
                    "format": "uuid",
                    "nullable": true,
                    "type": "string"
                  },
                  "panoramaAssetStatus": {
                    "enum": ["pending", "uploaded", "processing", "ready", "failed"],
                    "nullable": true,
                    "type": "string"
                  },
                  "sortOrder": {
                    "example": 0,
                    "type": "integer"
                  }
                },
                "required": ["destinationId", "name", "geoPoint"],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "altitude": {
                      "nullable": true,
                      "type": "number"
                    },
                    "createdAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "destinationId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "geoPoint": {
                      "properties": {
                        "latitude": {
                          "example": 18.3421,
                          "type": "number"
                        },
                        "longitude": {
                          "example": 105.9032,
                          "type": "number"
                        }
                      },
                      "required": ["latitude", "longitude"],
                      "type": "object"
                    },
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "initialFov": {
                      "type": "number"
                    },
                    "initialHeading": {
                      "type": "number"
                    },
                    "initialPitch": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "panoramaAssetId": {
                      "format": "uuid",
                      "nullable": true,
                      "type": "string"
                    },
                    "panoramaAssetStatus": {
                      "enum": ["pending", "uploaded", "processing", "ready", "failed"],
                      "nullable": true,
                      "type": "string"
                    },
                    "sortOrder": {
                      "type": "integer"
                    },
                    "status": {
                      "enum": ["draft", "published", "archived"],
                      "type": "string"
                    },
                    "updatedAt": {
                      "format": "date-time",
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "destinationId",
                    "name",
                    "geoPoint",
                    "altitude",
                    "panoramaAssetId",
                    "panoramaAssetStatus",
                    "initialHeading",
                    "initialPitch",
                    "initialFov",
                    "status",
                    "sortOrder",
                    "createdAt",
                    "updatedAt"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Created immersive scene draft."
          }
        },
        "summary": "",
        "tags": ["admin-immersive"]
      }
    },
    "/api/v1/admin/scenes/{id}": {
      "patch": {
        "operationId": "updateScene",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "altitude": {
                    "nullable": true,
                    "type": "number"
                  },
                  "destinationId": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "geoPoint": {
                    "properties": {
                      "latitude": {
                        "example": 18.3421,
                        "type": "number"
                      },
                      "longitude": {
                        "example": 105.9032,
                        "type": "number"
                      }
                    },
                    "required": ["latitude", "longitude"],
                    "type": "object"
                  },
                  "id": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "initialFov": {
                    "example": 90,
                    "type": "number"
                  },
                  "initialHeading": {
                    "example": 0,
                    "type": "number"
                  },
                  "initialPitch": {
                    "example": 0,
                    "type": "number"
                  },
                  "name": {
                    "example": "Cổng vào khu di tích",
                    "type": "string"
                  },
                  "panoramaAssetId": {
                    "format": "uuid",
                    "nullable": true,
                    "type": "string"
                  },
                  "panoramaAssetStatus": {
                    "enum": ["pending", "uploaded", "processing", "ready", "failed"],
                    "nullable": true,
                    "type": "string"
                  },
                  "sortOrder": {
                    "example": 0,
                    "type": "integer"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "altitude": {
                      "nullable": true,
                      "type": "number"
                    },
                    "createdAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "destinationId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "geoPoint": {
                      "properties": {
                        "latitude": {
                          "example": 18.3421,
                          "type": "number"
                        },
                        "longitude": {
                          "example": 105.9032,
                          "type": "number"
                        }
                      },
                      "required": ["latitude", "longitude"],
                      "type": "object"
                    },
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "initialFov": {
                      "type": "number"
                    },
                    "initialHeading": {
                      "type": "number"
                    },
                    "initialPitch": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "panoramaAssetId": {
                      "format": "uuid",
                      "nullable": true,
                      "type": "string"
                    },
                    "panoramaAssetStatus": {
                      "enum": ["pending", "uploaded", "processing", "ready", "failed"],
                      "nullable": true,
                      "type": "string"
                    },
                    "sortOrder": {
                      "type": "integer"
                    },
                    "status": {
                      "enum": ["draft", "published", "archived"],
                      "type": "string"
                    },
                    "updatedAt": {
                      "format": "date-time",
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "destinationId",
                    "name",
                    "geoPoint",
                    "altitude",
                    "panoramaAssetId",
                    "panoramaAssetStatus",
                    "initialHeading",
                    "initialPitch",
                    "initialFov",
                    "status",
                    "sortOrder",
                    "createdAt",
                    "updatedAt"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Updated immersive scene draft."
          }
        },
        "summary": "",
        "tags": ["admin-immersive"]
      }
    },
    "/api/v1/admin/scenes/{id}/publish": {
      "post": {
        "operationId": "publishScene",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "altitude": {
                      "nullable": true,
                      "type": "number"
                    },
                    "createdAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "destinationId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "geoPoint": {
                      "properties": {
                        "latitude": {
                          "example": 18.3421,
                          "type": "number"
                        },
                        "longitude": {
                          "example": 105.9032,
                          "type": "number"
                        }
                      },
                      "required": ["latitude", "longitude"],
                      "type": "object"
                    },
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "initialFov": {
                      "type": "number"
                    },
                    "initialHeading": {
                      "type": "number"
                    },
                    "initialPitch": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "panoramaAssetId": {
                      "format": "uuid",
                      "nullable": true,
                      "type": "string"
                    },
                    "panoramaAssetStatus": {
                      "enum": ["pending", "uploaded", "processing", "ready", "failed"],
                      "nullable": true,
                      "type": "string"
                    },
                    "sortOrder": {
                      "type": "integer"
                    },
                    "status": {
                      "enum": ["draft", "published", "archived"],
                      "type": "string"
                    },
                    "updatedAt": {
                      "format": "date-time",
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "destinationId",
                    "name",
                    "geoPoint",
                    "altitude",
                    "panoramaAssetId",
                    "panoramaAssetStatus",
                    "initialHeading",
                    "initialPitch",
                    "initialFov",
                    "status",
                    "sortOrder",
                    "createdAt",
                    "updatedAt"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Published immersive scene."
          }
        },
        "summary": "",
        "tags": ["admin-immersive"]
      }
    },
...
  },
  "servers": [],
...
}